msys2-runtime: Implement fixes for Wine
This is intended to address #682, and should eventually be upstreamed to
Cygwin to resolve https://bugs.winehq.org/show_bug.cgi?id=40528

0025: Do not directly use FastPebLock on the Wine code-path

Older versions of Wine did not populate this pointer, so instead call
the exported ntdll.dll function, which works correctly in all cases.

0026: Do not use FAST_CWD on Wine

The Cygwin cwd-handling code manipulates two internal Windows
structures, the Process Information Block and an undocumented global
dubbed FAST_CWD, in order to support some POSIX-like behaviours.

Wine doesn't have FAST_CWD, so disable that part of the manipulation
when Wine is detected.

0027: Unapplied logging change

This simply outputs some logs so that a crash or other bad behaviour can
be positively confirmed as being within or outside the CWD-override
support in Cygwin.

It should be dropped before applying, assuming everything else works.